+++ app/data_access/control_bulletin.py
@@ -23,4 +23,4 @@ class ControlBulletinFields(graphene.ObjectType):
     )
     business_type = graphene.String(required=False)
     vehicle_weight = graphene.String(required=False)
-    real_vehicle_weight = graphene.Int(required=False)
+    real_vehicle_weight = graphene.Float(required=False)
